(properly) protect against buffer overflow
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@18722 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/NetHandler.h
@@ -203,7 +203,7 @@ private:
   /// current TCP msg
   char tcpmsg[MaxPacketLen];
   /// bytes read in current msg
-  int tcplen;
+  size_t tcplen;
   /// current UDP msg
   static char	       udpmsg[MaxPacketLen];
   static int		udpLen;
